Add option to force icalEvent "alternative"
icalEvent adds the calendar event twice:
* once as a multipart/alternative,
* once as an attachment

This mimics the way Google Calendar sends invites, and works well with
most clients.

However, when there is a "regular" attachment, Nodemailer will not add
the multipart/alternative version, as it supposedly triggers a bug in
some clients.

This commit adds an option forceAlternative to the icalEvent object
which will force adding the multipart/alternative version even if there
is an attachment.